# Faster Bob.jar build