--- a/com.dynamo.cr/com.dynamo.cr.bob/gradle.properties
+++ b/com.dynamo.cr/com.dynamo.cr.bob/gradle.properties
@@ -1,0 +1,14 @@
+# Gradle performance optimizations
+org.gradle.parallel=true
+org.gradle.configureondemand=true
+org.gradle.daemon=true
+org.gradle.daemon.idletimeout=10800000
+org.gradle.jvmargs=-Xmx4g -Xms2g -XX:+UseG1GC -XX:MaxGCPauseMillis=200 -XX:+UseStringDeduplication -XX:+OptimizeStringConcat -XX:+UseCompressedOops -XX:+UseCompressedClassPointers -XX:+TieredCompilation -XX:TieredStopAtLevel=1 -XX:MaxMetaspaceSize=1g -XX:+HeapDumpOnOutOfMemoryError --add-opens=java.base/java.util=ALL-UNNAMED --add-opens=java.base/java.lang=ALL-UNNAMED --add-opens=java.base/java.lang.invoke=ALL-UNNAMED --add-opens=java.prefs/java.util.prefs=ALL-UNNAMED --add-opens=java.base/java.nio.charset=ALL-UNNAMED --add-opens=java.base/java.net=ALL-UNNAMED --add-opens=java.base/java.util.concurrent.atomic=ALL-UNNAMED
+
+# Java compilation optimizations
+org.gradle.java.compile.max-workers=8
+org.gradle.workers.max=8
+
+# File system optimizations
+org.gradle.vfs.watch=true
+org.gradle.vfs.verbose=false