15%-30% faster bob build (depends on hardware).

CI win 32:
Before
bob_light                     51.33 s    51.33 s    51.33 s

After
bob_light                     39.35 s    39.35 s    39.35 s

CI x86_64_osx:

Before 
bob_light          31.20 s               31.20 s    31.20 s

After
bob_light          25.46 s               25.46 s    25.46 s


CI build_bob:

Before:
BUILD SUCCESSFUL in 1m 3s

After:
BUILD SUCCESSFUL in 51s